requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
google-generativeai
streamlit-chat
openpyxl
altair==5.2.0
joblib
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.2.5
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy>=1.26.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.34.0
requests==2.32.3
rich==13.7.1   # ✅ compatible with streamlit 1.38.0
rpds-py>=0.20.0
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sklearn-pandas==2.2.0
smmap==5.0.1
streamlit==1.19
streamlit-chat==0.1.1
tenacity==8.2.3
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
Werkzeug==3.0.3
google-generativeai==0.8.5

# Build helpers
setuptools>=65.5.0
wheel
